move Request from [org.matsim.contrib.dvrp].data to .optimizer
