+++ contribs/dvrp/src/main/java/org/matsim/contrib/dvrp/passenger/PassengerRequest.java
@@ -22,7 +22,7 @@ package org.matsim.contrib.dvrp.passenger;
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.population.Person;
-import org.matsim.contrib.dvrp.data.Request;
+import org.matsim.contrib.dvrp.optimizer.Request;
 
 public interface PassengerRequest extends Request {
 	/**
