+++ contribs/dvrp/src/main/java/org/matsim/contrib/dvrp/passenger/PassengerRequestCreator.java
@@ -21,7 +21,7 @@ package org.matsim.contrib.dvrp.passenger;
 
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.network.Link;
-import org.matsim.contrib.dvrp.data.Request;
+import org.matsim.contrib.dvrp.optimizer.Request;
 import org.matsim.core.mobsim.framework.MobsimPassengerAgent;
 
 /**
