+++ contribs/taxi/src/main/java/org/matsim/contrib/taxi/optimizer/TaxiOptimizationValidation.java
@@ -21,7 +21,7 @@ package org.matsim.contrib.taxi.optimizer;
 
 import java.util.stream.Stream;
 
-import org.matsim.contrib.dvrp.data.Request;
+import org.matsim.contrib.dvrp.optimizer.Request;
 import org.matsim.contrib.dvrp.fleet.Fleet;
 import org.matsim.contrib.taxi.passenger.TaxiRequest.TaxiRequestStatus;
 import org.matsim.contrib.taxi.passenger.TaxiRequests;
