+++ contribs/taxi/src/main/java/org/matsim/contrib/taxi/schedule/reconstruct/RequestRecorder.java
@@ -29,7 +29,7 @@ import org.matsim.api.core.v01.events.handler.PersonDepartureEventHandler;
 import org.matsim.api.core.v01.events.handler.PersonEntersVehicleEventHandler;
 import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.population.Person;
-import org.matsim.contrib.dvrp.data.Request;
+import org.matsim.contrib.dvrp.optimizer.Request;
 import org.matsim.contrib.taxi.passenger.TaxiRequest;
 
 public class RequestRecorder implements PersonDepartureEventHandler, PersonEntersVehicleEventHandler {
