+++ contribs/taxi/src/main/java/org/matsim/contrib/taxi/schedule/reconstruct/ScheduleReconstructor.java
@@ -26,7 +26,7 @@ import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.network.Network;
 import org.matsim.api.core.v01.population.Person;
-import org.matsim.contrib.dvrp.data.Request;
+import org.matsim.contrib.dvrp.optimizer.Request;
 import org.matsim.contrib.dvrp.fleet.Fleet;
 import org.matsim.contrib.dvrp.fleet.FleetImpl;
 import org.matsim.contrib.taxi.passenger.TaxiRequest;
