+++ contribs/taxi/src/main/java/org/matsim/contrib/taxi/schedule/reconstruct/TaxiRequestWithModifiableToLink.java
@@ -21,7 +21,7 @@ package org.matsim.contrib.taxi.schedule.reconstruct;
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.population.Person;
-import org.matsim.contrib.dvrp.data.Request;
+import org.matsim.contrib.dvrp.optimizer.Request;
 import org.matsim.contrib.taxi.passenger.TaxiRequest;
 
 /**
